+++ cmem_plugin_reason/utils.py
@@ -97,18 +97,22 @@ def create_xml_catalog_file(dir_: str, graphs: dict) -> None:
 
 def get_graphs_tree(plugin: WorkflowPlugin, graph_iris: tuple, ignore_missing: bool = True) -> dict:
     """Get graph import tree. Last item in graph_iris is output_graph_iris which is excluded"""
+    graphs_list = [_["iri"] for _ in get_graphs_list()]
+    missing = []
     graphs = {}
     for graph_iri in graph_iris[:-1]:
-        if graph_iri not in graphs:
+        if graph_iri not in graphs_list:
+            missing.append(graph_iri)
+        elif graph_iri not in graphs:
             graphs[graph_iri] = f"{token_hex(8)}.nt"
             tree = get_graph_import_tree(graph_iri)
             for value in tree["tree"].values():
                 for iri in value:
-                    if iri not in graphs and iri != graph_iri[-1]:
-                        graphs[iri] = f"{token_hex(8)}.nt"
-
-    graphs_list = [_["iri"] for _ in get_graphs_list()]
-    missing = [_ for _ in graphs if _ not in graphs_list]
+                    if iri != graph_iris[-1]:
+                        if iri not in graphs_list:
+                            missing.append(iri)
+                        elif iri not in graphs:
+                            graphs[iri] = f"{token_hex(8)}.nt"
     if ignore_missing:
         [plugin.log.warning(f"Missing graph import: {iri}") for iri in missing]
     else: